+++ src/TrendyolSharp.Marketplace/Models/Base/ClaimItemReason.cs
@@ -3,6 +3,6 @@
 public sealed class ClaimItemReason
 {
   public string Name { get; set; }
-  public int ExternalReasonId { get; set; }
+  public long ExternalReasonId { get; set; }
   public string Code { get; set; }
 }
