+++ src/TrendyolSharp.Marketplace/Models/Base/CreateClaimItem.cs
@@ -4,6 +4,6 @@ public sealed class CreateClaimItem
 {
   public string Barcode { get; set; }
   public string CustomerNote { get; set; }
-  public int Quantity { get; set; }
-  public int ReasonId { get; set; }
+  public long Quantity { get; set; }
+  public long ReasonId { get; set; }
 }
