+++ src/TrendyolSharp.Marketplace/Models/Base/PackageDetail.cs
@@ -3,5 +3,5 @@
 public sealed class PackageDetail
 {
   public long OrderLineId { get; set; }
-  public int Quantities { get; set; }
+  public long Quantities { get; set; }
 }
