+++ src/TrendyolSharp.Marketplace/Models/Base/ProductsAttribute.cs
@@ -2,8 +2,8 @@
 
 public sealed class ProductsAttribute
 {
-  public int AttributeId { get; set; }
+  public long AttributeId { get; set; }
   public string AttributeName { get; set; }
-  public int? AttributeValueId { get; set; }
+  public long? AttributeValueId { get; set; }
   public string AttributeValue { get; set; }
 }
