+++ src/TrendyolSharp.Marketplace/Models/Base/ShipmentAddress.cs
@@ -10,11 +10,11 @@ public sealed class ShipmentAddress
   public string Address2 { get; set; }
   public string City { get; set; }
   public string District { get; set; }
-  public int DistrictId { get; set; }
-  public int CityCode { get; set; }
+  public long DistrictId { get; set; }
+  public long CityCode { get; set; }
   public string PostalCode { get; set; }
   public string CountryCode { get; set; }
-  public int NeighborhoodId { get; set; }
+  public long NeighborhoodId { get; set; }
   public string Neighborhood { get; set; }
   public string Phone { get; set; }
   public string FullName { get; set; }
