+++ src/TrendyolSharp.Marketplace/Models/Request/RequestGetClaims.cs
@@ -4,8 +4,8 @@ public sealed class RequestGetClaims
 {
   public string? ClaimIds { get; set; }
   public string? ClaimItemStatus { get; set; }
-  public int? EndDate { get; set; }
-  public int? StartDate { get; set; }
+  public long? EndDate { get; set; }
+  public long? StartDate { get; set; }
   public string? OrderNumber { get; set; }
   public int? Size { get; set; }
   public int? Page { get; set; }
